# openapi-schema.yml
openapi: 3.0.3
info:
  title: NoCode Backend API
  version: 1.0.0
  description: API pour la plateforme NoCode - Création d'applications sans code
paths:
...
  /api/v1/automation/workflows/:
    get:
      operationId: automation_workflows_list
      description: |-
        API pour gérer les workflows.

        Endpoints:
        - GET /api/automation/workflows/ : Liste des workflows
        - POST /api/automation/workflows/ : Créer un workflow
        - GET /api/automation/workflows/{id}/ : Détails d'un workflow
        - PUT /api/automation/workflows/{id}/ : Mettre à jour un workflow
        - DELETE /api/automation/workflows/{id}/ : Supprimer un workflow
        - POST /api/automation/workflows/{id}/execute/ : Exécuter un workflow
        - GET /api/automation/workflows/{id}/executions/ : Historique d'exécutions
        - POST /api/automation/workflows/{id}/validate/ : Valider un workflow
        - POST /api/automation/workflows/{id}/duplicate/ : Dupliquer un workflow
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - automation
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWorkflowListList'
          description: ''
    post:
      operationId: automation_workflows_create
      description: |-
        API pour gérer les workflows.

        Endpoints:
        - GET /api/automation/workflows/ : Liste des workflows
        - POST /api/automation/workflows/ : Créer un workflow
        - GET /api/automation/workflows/{id}/ : Détails d'un workflow
        - PUT /api/automation/workflows/{id}/ : Mettre à jour un workflow
        - DELETE /api/automation/workflows/{id}/ : Supprimer un workflow
        - POST /api/automation/workflows/{id}/execute/ : Exécuter un workflow
        - GET /api/automation/workflows/{id}/executions/ : Historique d'exécutions
        - POST /api/automation/workflows/{id}/validate/ : Valider un workflow
        - POST /api/automation/workflows/{id}/duplicate/ : Dupliquer un workflow
      tags:
      - automation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/WorkflowCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/WorkflowCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowCreate'
          description: ''
...
  /api/v1/automation/workflows/{id}/edges/:
    get:
      operationId: automation_workflows_edges_list
      description: ViewSet pour les opérations CRUD sur les arêtes.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - automation
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEdgeList'
          description: ''
    post:
      operationId: automation_workflows_edges_create
      description: ViewSet pour les opérations CRUD sur les arêtes.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - automation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EdgeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EdgeRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edge'
          description: ''
...
  /api/v1/automation/workflows/{id}/nodes/:
    get:
      operationId: automation_workflows_nodes_list
      description: ViewSet pour les opérations CRUD sur les nœuds.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - automation
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNodeList'
          description: ''
    post:
      operationId: automation_workflows_nodes_create
      description: ViewSet pour les opérations CRUD sur les nœuds.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - automation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NodeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/NodeRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
          description: ''
...
  /api/v1/runtime/apps/:
    get:
      operationId: runtime_apps_list
      description: Retourne la liste des applications générées par l'utilisateur courant.
      summary: Lister les applications générées
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - runtime
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedGeneratedAppList'
          description: ''
        '403':
          description: Accès non autorisé
    post:
      operationId: runtime_apps_create
      description: Crée une nouvelle application générée.
      summary: Créer une nouvelle application
      tags:
      - runtime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratedAppRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GeneratedAppRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GeneratedAppRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedApp'
          description: ''
        '400':
          description: Données invalides
        '403':
          description: Accès non autorisé
...
  /api/v1/runtime/apps/{id}/deploy/:
    post:
      operationId: runtime_apps_deploy_create
      description: Lance le déploiement d'une application générée sur l'environnement
        cible.
      summary: Déployer une application
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this Application générée.
        required: true
      tags:
      - runtime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentActionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DeploymentActionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DeploymentActionRequest'
      security:
      - jwtAuth: []
      responses:
        '202':
          description: Déploiement démarré avec succès
        '400':
          description: L'application n'est pas dans un état déployable
        '403':
          description: Accès non autorisé
        '404':
          description: Application non trouvée
        '500':
          description: Erreur lors du démarrage du déploiement
...
  /api/v1/studio/projects/:
    get:
      operationId: studio_projects_list
      description: "\n        **Récupère la liste de tous les projets accessibles\
        \ par l'utilisateur connecté.**\n        \n        ## \U0001F3AF Projets retournés\
        \ :\n        \n        ### Pour un Client Individuel (sans organisation) :\n\
        \        - ✅ Tous les projets qu'il a créés personnellement\n        - ✅ Les\
        \ projets où `organization` est NULL\n        \n        ### Pour un Membre\
        \ d'Organisation :\n        - ✅ Tous ses projets personnels (créés par lui,\
        \ organization=NULL)\n        - ✅ Tous les projets de ses organisations (où\
        \ il est membre actif)\n        - ✅ Visible par tous les membres (OWNER, ADMIN,\
        \ MEMBER)\n        \n        ## \U0001F4CA Réponse :\n        - Liste de projets\
        \ avec leurs détails complets\n        - Inclut : tracking_id, name, schema_name,\
        \ organization, created_by, dates\n        \n        ## \U0001F510 Permissions\
        \ :\n        - Authentification requise (Bearer Token)\n        - Retourne\
        \ uniquement les projets accessibles par l'utilisateur\n        \n       \
        \ ## \U0001F4A1 Exemple d'utilisation :\n        ```\n        GET /api/v1/studio/projects/\n\
        \        Authorization: Bearer <votre_token>\n        ```\n        "
      summary: 📋 Lister tous les projets accessibles
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - 📁 Studio - Projets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjectList'
          description: ''
    post:
      operationId: studio_projects_create
      description: "\n        **Crée un nouveau projet NoCode.**\n        \n     \
        \   ## \U0001F3AF Types de projets :\n        \n        ### 1️⃣ Projet Personnel\
        \ (Client Individuel) :\n        ```json\n        {\n            \"name\"\
        : \"Mon Application Mobile\"\n            // Ne pas fournir organization_id\n\
        \        }\n        ```\n        - Le projet sera lié uniquement au créateur\n\
        \        - `organization` sera NULL\n        - Seul le créateur aura accès\n\
        \        \n        ### 2️⃣ Projet d'Organisation :\n        ```json\n    \
        \    {\n            \"name\": \"ERP Entreprise\",\n            \"organization_id\"\
        : \"uuid-de-votre-organisation\"\n        }\n        ```\n        - Le projet\
        \ sera lié à l'organisation\n        - Tous les membres actifs de l'organisation\
        \ auront accès\n        - OWNER/ADMIN peuvent modifier, MEMBER peut lire\n\
        \        \n        ## ⚙️ Ce qui est créé automatiquement :\n        - ✅ Un\
        \ schéma PostgreSQL unique (`schema_name`)\n        - ✅ Une page d'accueil\
        \ par défaut\n        - ✅ Une structure de base pour l'application\n     \
        \   \n        ## \U0001F510 Permissions :\n        - Authentification requise\n\
        \        - Pour créer un projet d'organisation : être membre actif de cette\
        \ organisation\n        \n        ## \U0001F4CA Réponse :\n        - Détails\
        \ complets du projet créé\n        - Code HTTP 201 Created\n        \n   \
        \     ## ⚠️ Notes importantes :\n        - Le nom du projet doit être unique\
        \ dans votre espace\n        - Le schema_name est généré automatiquement\n\
        \        - La suppression d'un projet supprime toutes ses données\n      \
        \  "
      summary: ➕ Créer un nouveau projet
      tags:
      - 📁 Studio - Projets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
            examples:
              ProjetPersonnel:
                value:
                  name: Mon Portfolio
                summary: Projet Personnel
              ProjetOrganisation:
                value:
                  name: CRM Enterprise
                  organization_id: 550e8400-e29b-41d4-a716-446655440000
                summary: Projet Organisation
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
...
  /api/v1/studio/projects/{id}/schemas/:
    get:
      operationId: studio_projects_schemas_retrieve
      description: Liste des schémas de données du projet
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - studio
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
...
components:
  schemas:
...
    Edge:
      type: object
      description: Serializer pour les arêtes du graphe de workflows.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        workflow:
          type: string
          format: uuid
        source_node:
          type: string
          format: uuid
        target_node:
          type: string
          format: uuid
        source_port:
          type: string
          default: output
          maxLength: 50
        target_port:
          type: string
          default: input
          maxLength: 50
        label:
          type: string
          maxLength: 255
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - source_node
      - target_node
      - workflow
    EdgeRequest:
      type: object
      description: Serializer pour les arêtes du graphe de workflows.
      properties:
        workflow:
          type: string
          format: uuid
        source_node:
          type: string
          format: uuid
        target_node:
          type: string
          format: uuid
        source_port:
          type: string
          minLength: 1
          default: output
          maxLength: 50
        target_port:
          type: string
          minLength: 1
          default: input
          maxLength: 50
        label:
          type: string
          maxLength: 255
      required:
      - source_node
      - target_node
      - workflow
...
    Node:
      type: object
      description: Serializer pour les nœuds du graphe de workflows.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        workflow:
          type: string
          format: uuid
        node_id:
          type: string
          maxLength: 100
        node_type:
          enum:
          - trigger
          - action
          - condition
          - math
          - string
          - data
          - api
          - email
          type: string
          description: |-
            * `trigger` - Déclencheur
            * `action` - Action
            * `condition` - Condition
            * `math` - Math
            * `string` - String
            * `data` - Data
            * `api` - API
            * `email` - Email
        label:
          type: string
          maxLength: 255
        position_x:
          type: number
          format: double
        position_y:
          type: number
          format: double
        config:
          type: object
          additionalProperties: {}
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - label
      - node_id
      - node_type
      - workflow
    NodeRequest:
      type: object
      description: Serializer pour les nœuds du graphe de workflows.
      properties:
        workflow:
          type: string
          format: uuid
        node_id:
          type: string
          minLength: 1
          maxLength: 100
        node_type:
          enum:
          - trigger
          - action
          - condition
          - math
          - string
          - data
          - api
          - email
          type: string
          description: |-
            * `trigger` - Déclencheur
            * `action` - Action
            * `condition` - Condition
            * `math` - Math
            * `string` - String
            * `data` - Data
            * `api` - API
            * `email` - Email
        label:
          type: string
          minLength: 1
          maxLength: 255
        position_x:
          type: number
          format: double
        position_y:
          type: number
          format: double
        config:
          type: object
          additionalProperties: {}
      required:
      - label
      - node_id
      - node_type
      - workflow
...
    PaginatedEdgeList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Edge'
    PaginatedGeneratedAppList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/GeneratedApp'
...
    PaginatedNodeList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Node'
...
    PaginatedWorkflowListList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowList'
...
    WorkflowCreate:
      type: object
      description: Serializer pour créer un workflow.
      properties:
        name:
          type: string
          title: Nom du workflow
          maxLength: 255
        description:
          type: string
        organization:
          type: integer
          title: Organisation
        project:
          type: integer
          nullable: true
          title: Projet
        status:
          enum:
          - DRAFT
          - ACTIVE
          - PAUSED
          - ARCHIVED
          type: string
          description: |-
            * `DRAFT` - Brouillon
            * `ACTIVE` - Actif
            * `PAUSED` - En pause
            * `ARCHIVED` - Archivé
          title: Statut
        config:
          type: object
          additionalProperties: {}
          title: Configuration
          description: Configuration générale du workflow
        variables:
          type: object
          additionalProperties: {}
          description: Variables globales du workflow
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowStep'
        triggers:
          type: array
          items:
            $ref: '#/components/schemas/Trigger'
      required:
      - name
      - organization
    WorkflowCreateRequest:
      type: object
      description: Serializer pour créer un workflow.
      properties:
        name:
          type: string
          minLength: 1
          title: Nom du workflow
          maxLength: 255
        description:
          type: string
        organization:
          type: integer
          title: Organisation
        project:
          type: integer
          nullable: true
          title: Projet
        status:
          enum:
          - DRAFT
          - ACTIVE
          - PAUSED
          - ARCHIVED
          type: string
          description: |-
            * `DRAFT` - Brouillon
            * `ACTIVE` - Actif
            * `PAUSED` - En pause
            * `ARCHIVED` - Archivé
          title: Statut
        config:
          type: object
          additionalProperties: {}
          title: Configuration
          description: Configuration générale du workflow
        variables:
          type: object
          additionalProperties: {}
          description: Variables globales du workflow
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowStepRequest'
        triggers:
          type: array
          items:
            $ref: '#/components/schemas/TriggerRequest'
      required:
      - name
      - organization
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: http://localhost:8000
  description: Development server
- url: https://api.nocode.com
  description: Production server